Move model to models package
