+++ src/main/scala/com/gsoc/models/Degree.scala
@@ -1,3 +1,3 @@
-package com.gsoc.gremlin_graph
+package com.gsoc.models
 
 final case class Degree(in: Long, out: Long, overall: Long)
